// src/types.d.ts
    export interface TemplateConfig {
        name            : TemplateVariant;
        label           : string;
        description     : string;
        repo            : string;
        deps?           : string[];
        requiresSetup   : boolean;
        ready           : boolean;
    }

    export interface SpaceTypeConfig {
        type            : SpaceType;
        defaultTemplate : TemplateVariant;
        templates       : TemplateConfig[];
    }

    export type SpaceType       = 'lib' | 'cli' | 'server' | 'app';
    export type TemplateVariant = 'clean';